api-django/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
Pillow==10.2.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
requests==2.31.0
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
pytz==2024.1
gunicorn==21.2.0

